requirements.txt
requests
beautifulsoup4
pyotp
python-dotenv
